Tidy up kernel and acqf getters in Inference wrapper. Removed redundant config parameter for BL dynamics model. Update configs for BO paper
