+++ input/inference/example.toml
@@ -90,7 +90,7 @@ config_version = "3.0"
 
         [params.stop.solid]                  # solidification criterion
             enabled           = true
-            phi_crit          = 0.01         # stop when global melt fraction < phi_crit
+            phi_crit          = 0.05         # stop when global melt fraction < phi_crit
             freeze_volatiles  = false        # freeze mantle volatiles at solidification
 
         [params.stop.radeqm]                 # radiative equilibrium criterion
@@ -117,7 +117,7 @@ config_version = "3.0"
     # THE_PLANETS
     # mass_tot     = 3.0   # SN
     #  mass_tot     = 2.0   # SE
-    mass_tot     = 1.2   # TR
+    mass_tot     = 1.1   # TR
 
     # Initial temperature profile
     temperature_mode          = "adiabatic"  # default; other modes: isothermal | linear | adiabatic | adiabatic_from_cmb | accretion | isentropic
@@ -157,7 +157,7 @@ config_version = "3.0"
         # THE_PLANETS
         # C_budget      = 0.5       # SN
         #  C_budget      = 1.0     # SE
-        C_budget        = 3.0     # TR
+        C_budget        = 2.0     # TR
 
 
         N_mode                = "N/H"        # N/H mass ratio | ppmw | kg
@@ -168,7 +168,7 @@ config_version = "3.0"
         # THE_PLANETS
         # S_budget      = 0.8      # SN
         #  S_budget      = 3.0    # SE
-        S_budget      = 2.0    # TR
+        S_budget      = 1.5    # TR
 
         use_metallicity       = false        # scale C/N/S from solar metallicity; replaces C/N/S budgets
         metallicity           = 1000         # metallicity relative to solar [by mass]
@@ -434,9 +434,8 @@ config_version = "3.0"
         thermal_conductivity      = 4.2          # thermal conductivity [W/m/K]
         thermal_diffusivity       = 1.0e-6       # thermal diffusivity [m^2/s]
         thermal_expansivity       = 2.0e-5       # thermal expansivity [1/K]
-        viscosity_model           = 2            # 1=constant | 2=aggregate smooth | 3=Arrhenius
+        viscosity_model           = 3            # 1=constant | 2=aggregate smooth | 3=Arrhenius
         eta_constant              = 1.0e2        # constant viscosity for model 1 [Pa s]
-        transition_width          = 0.15          # width of viscosity transition (melt-fraction units)
         eta_solid_const           = 5.0e22       # solid-end viscosity for aggregate model [Pa s]
         eta_melt_const            = 1.0e2        # melt-end viscosity for aggregate model [Pa s]
         dynamic_viscosity         = 3.8e9        # Arrhenius solid reference dynamic viscosity [Pa s]
@@ -542,7 +541,7 @@ config_version = "3.0"
         latent_heat           = false        # latent heat from condensation/evaporation
         real_gas              = false        # real-gas EOS where available
         chemistry             = "none"       # atmospheric chemistry: none | eq (FastChem)
-        thermo_functions      = true        # use T-dependent thermo functions
+        thermo_functions      = false        # use T-dependent thermo functions
         mlt_criterion         = "s"          # convection criterion: s (Schwarzschild) | l (Ledoux)
 
         # Surface properties
